Add moveDown, moveLeft and freeze functions.
Add moveDown() function so that the tetrominoes are moved
down each second. Along with it, the freeze function will
check, for every second, if the current tetromino is located
at the bottom of the screen or blocked by another tetromino,
in which case it will freeze its movement.

The moveLeft function works so that the tetromino is able to
move to its left, unless it is at the edge of our playable screen,
in which case it'll have its position reset.

Signed-off-by: Victor <victorcampos@usp.br>