Add start/pause button functionality.
Add functionality to the start/pause game button by freezing the game
if pressed when it is already running.
Comment original timerId variable line so that the game only starts
when the start button is pressed.

Fix displaySquares const so as to add them to .mini-grid div
class, thus fixing the "next up tetromino" functionality.

Signed-off-by: Victor <victorcampos@usp.br>