Add addScore function to game.
Add addScore() function, which defines a for loop and checks
the row for completefulness. if row is full, +10 is added to
the player's score and the squares are removed from the 'tetromino'
class.
Add comment to fixme on function rotate() as it may glitch if player
attempts to rotate at the edge of the screen.

Signed-off-by: Victor <victorcampos@usp.br>